requirements.txt
contourpy==1.3.1
cycler==0.12.1
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.3.0
imbalanced-learn==0.13.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.0
six==1.17.0
sklearn-compat==0.1.3
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
wheel==0.45.1
xgboost==3.0.0
